refactor remove instance actions into separate mixin to be shared across components
